+++ src/global.d.ts
@@ -7,3 +7,7 @@ interface CSSStyleDeclaration {
 interface DocumentType extends Node, ChildNode {
     readonly internalSubset: string | null;
 }
+
+interface Document {
+    fonts: any;
+}